Make the calculator work, more or less
I think I have something that will end up working pretty nicely. It
needs to be cleaned up a bit and expanded, but I want to make sure I
save my progress while it's *sort of* working before I mess anything up.

Creating an object to hold all of the calculator data was a real
game-changer for this.